requirements.txt
alembic==1.7.6
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
importlib-metadata==4.11.1
importlib-resources==5.4.0
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
python-dotenv==0.19.2
SQLAlchemy==1.4.31
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.7.0
